requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-chroma>=0.1.0
chromadb>=0.4.20
pydantic>=2.5.0
pydantic-settings>=2.1.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.25.0
python-multipart>=0.0.6
aiofiles>=23.2.1

# Document processing
pdfminer.six>=20231228
python-docx>=1.1.0
openpyxl>=3.1.2
python-pptx>=0.6.23
markdown>=3.5.1
beautifulsoup4>=4.12.2
pypdf>=3.17.4
unstructured>=0.11.0
pytesseract>=0.3.10
pillow>=10.1.0

# LLM providers
openai>=1.10.0
dashscope>=1.14.0
zhipuai>=1.0.0
erniebot>=0.5.0

# Vector databases
pymilvus>=2.3.0
pinecone-client>=2.2.4
qdrant-client>=1.6.0
weaviate-client>=3.24.0

# Utilities
httpx>=0.25.0
tenacity>=8.2.0
tiktoken>=0.5.0
sentence-transformers>=2.2.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6